css/css-grid-1/alignment/grid-alignment-implies-size-change-034.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Changes on Default-Alignment may affect grid item's width</title>
<link rel="author" title="Javier Fernandez Garcia-Boente" href="mailto:jfernandez@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid/#row-align">
<link rel="help" href="https://drafts.csswg.org/css-grid/#grid-item-sizing">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-justify-items">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-justify-self-start">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-justify-self-normal">
<meta name="assert" content="Changing the grid item's justify-items value from 'normal' to 'start' causes no effect on replaced items.">
<style>
.grid {
  position: relative;
  display: inline-grid;
  grid-template-columns: 200px;
  grid-template-rows: 100px;
  font: 20px/1 Ahem;
  background: grey;
  justify-items: normal;
}
#item {
  background: blue;
  justify-self: auto;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/check-layout-th.js"></script>
<script src="../support/style-change.js"></script>
<div class="grid" id="grid">
  <img data-expected-height="100" id="item" src="../grid-items/support/100x100-green.png"></img>
</div>
<script>
  document.body.offsetLeft;
  evaluateStyleChange(item, "before", "data-expected-width", 100);
  grid.style.justifyItems = "start";
  evaluateStyleChange(item, "after", "data-expected-width", 100);
</script>
